# Performance Improvements for files with thousands of tests