--- a/lib/enhance-assert.js
+++ b/lib/enhance-assert.js
@@ -30,7 +30,8 @@ function enhanceAssert(opts) {
 			onError: opts.onError,
 			onSuccess: opts.onSuccess,
 			patterns: module.exports.PATTERNS,
-			wrapOnlyPatterns: module.exports.NON_ENHANCED_PATTERNS
+			wrapOnlyPatterns: module.exports.NON_ENHANCED_PATTERNS,
+			bindReceiver: false
 		}
 	);