--- a/lib/runner.js
+++ b/lib/runner.js
@@ -49,6 +49,7 @@ function Runner(opts) {
 	this.options = opts || {};
 	this.results = [];
 	this.tests = [];
+	this.testsByType = {};
 }
 
 util.inherits(Runner, EventEmitter);
@@ -58,8 +59,15 @@ optionChain(chainableMethods, function (opts, title, fn) {
 	var Constructor = (opts && /Each/.test(opts.type)) ? Hook : Test;
 	var test = new Constructor(title, fn);
 	test.metadata = objectAssign({}, opts);
-	this.tests.push(test);
+	this._addTest(test);
 }, Runner.prototype);
+
+Runner.prototype._addTest = function (test) {
+	this.tests.push(test);
+	var type = test.metadata.type;
+	var tests = this.testsByType[type] || (this.testsByType[type] = []);
+	tests.push(test);
+};
 
 Runner.prototype._runTestWithHooks = function (test) {
 	if (test.metadata.skipped) {
@@ -70,9 +78,9 @@ Runner.prototype._runTestWithHooks = function (test) {
 		return hook.test(test.title);
 	}
 
-	var tests = this.select({type: 'beforeEach'}).map(hookToTest);
+	var tests = (this.testsByType.beforeEach || []).map(hookToTest);
 	tests.push(test);
-	tests.push.apply(tests, this.select({type: 'afterEach'}).map(hookToTest));
+	tests.push.apply(tests, (this.testsByType.afterEach || []).map(hookToTest));
 
 	var context = {};
 
@@ -188,7 +196,9 @@ Runner.prototype.run = function () {
 };
 
 Runner.prototype.select = function (filter) {
-	return this.tests.filter(function (test) {
+	var tests = filter.type ? this.testsByType[filter.type] || [] : this.tests;
+
+	return tests.filter(function (test) {
 		return Object.keys(filter).every(function (key) {
 			return filter[key] === test.metadata[key];
 		});