--- a/lib/test.js
+++ b/lib/test.js
@@ -30,13 +30,6 @@ function Test(title, fn) {
 	this.duration = null;
 	this.assertError = undefined;
 
-	Object.defineProperty(this, 'assertCount', {
-		enumerable: true,
-		get: function () {
-			return this.assertions.length;
-		}
-	});
-
 	// TODO(jamestalmage): make this an optional constructor arg instead of having Runner set it after the fact.
 	// metadata should just always exist, otherwise it requires a bunch of ugly checks all over the place.
 	this.metadata = {};
@@ -52,6 +45,13 @@ function Test(title, fn) {
 }
 
 module.exports = Test;
+
+Object.defineProperty(Test.prototype, 'assertCount', {
+	enumerable: true,
+	get: function () {
+		return this.assertions.length;
+	}
+});
 
 Test.prototype._assert = function (promise) {
 	this.assertions.push(promise);