--- a/lib/test.js
+++ b/lib/test.js
@@ -223,71 +223,73 @@ Test.prototype.exit = function () {
 };
 
 Test.prototype._publicApi = function () {
-	var self = this;
-	var api = {skip: {}};
-
-	// Getters
-	[
-		'assertCount',
-		'title',
-		'end'
-	]
-		.forEach(function (name) {
-			Object.defineProperty(api, name, {
-				enumerable: name === 'end' ? self.metadata.callback : true,
-				get: function () {
-					return self[name];
-				}
-			});
-		});
-
-	// Get / Set
-	Object.defineProperty(api, 'context', {
-		enumerable: true,
-		get: function () {
-			return self.context;
-		},
-		set: function (context) {
-			self.context = context;
-		}
-	});
-
-	// Bound Functions
-	api.plan = this.plan.bind(this);
-
-	function skipFn() {
-		self._assert(Promise.resolve());
-	}
-
-	function onAssertionEvent(event) {
-		var promise;
-		if (event.assertionThrew) {
-			event.error.powerAssertContext = event.powerAssertContext;
-			promise = Promise.reject(event.error);
-		} else {
-			var ret = event.returnValue;
-			promise = isObservable(ret) ? observableToPromise(ret) : Promise.resolve(ret);
-		}
-		promise = promise
-			.catch(function (err) {
-				err.originalMessage = event.originalMessage;
-				return Promise.reject(err);
-			});
-		self._assert(promise);
-		return promise;
-	}
-
-	var enhanced = enhanceAssert({
-		assert: assert,
-		onSuccess: onAssertionEvent,
-		onError: onAssertionEvent
-	});
-
-	// Patched assert methods: increase assert count and store errors.
-	Object.keys(assert).forEach(function (el) {
-		api.skip[el] = skipFn;
-		api[el] = enhanced[el].bind(enhanced);
-	});
-
-	return api;
+	return new PublicApi(this);
 };
+
+function PublicApi(test) {
+	this._test = test;
+	this.plan = test.plan.bind(test);
+	this.skip = new SkipApi(test);
+}
+
+function onAssertionEvent(event) {
+	var promise;
+	if (event.assertionThrew) {
+		event.error.powerAssertContext = event.powerAssertContext;
+		promise = Promise.reject(event.error);
+	} else {
+		var ret = event.returnValue;
+		promise = isObservable(ret) ? observableToPromise(ret) : Promise.resolve(ret);
+	}
+	promise = promise
+		.catch(function (err) {
+			err.originalMessage = event.originalMessage;
+			return Promise.reject(err);
+		});
+	this._test._assert(promise);
+	return promise;
+}
+
+PublicApi.prototype = enhanceAssert({
+	assert: assert,
+	onSuccess: onAssertionEvent,
+	onError: onAssertionEvent
+});
+
+// Getters
+[
+	'assertCount',
+	'title',
+	'end'
+]
+	.forEach(function (name) {
+		Object.defineProperty(PublicApi.prototype, name, {
+			enumerable: false,
+			get: function () {
+				return this._test[name];
+			}
+		});
+	});
+
+// Get / Set
+Object.defineProperty(PublicApi.prototype, 'context', {
+	enumerable: true,
+	get: function () {
+		return this._test.context;
+	},
+	set: function (context) {
+		this._test.context = context;
+	}
+});
+
+function skipFn() {
+	return this._test._assert(Promise.resolve());
+}
+
+function SkipApi(test) {
+	this._test = test;
+}
+
+Object.keys(assert).forEach(function (el) {
+	SkipApi.prototype[el] = skipFn;
+});