--- a/lib/test.js
+++ b/lib/test.js
@@ -81,6 +81,17 @@ Test.prototype.plan = function (count) {
 	this.planStack = new Error().stack;
 };
 
+Test.prototype._run = function () {
+	var ret;
+	try {
+		ret = this.fn(this._publicApi());
+	} catch (err) {
+		this._setAssertError(err);
+		this.exit();
+	}
+	return ret;
+};
+
 Test.prototype.run = function () {
 	var self = this;
 
@@ -101,14 +112,7 @@ Test.prototype.run = function () {
 	// wait until all assertions are complete
 	this._timeout = globals.setTimeout(function () {}, maxTimeout);
 
-	var ret;
-
-	try {
-		ret = this.fn(this._publicApi());
-	} catch (err) {
-		this._setAssertError(err);
-		this.exit();
-	}
+	var ret = this._run();
 
 	var asyncType = 'promises';