--- a/lib/test.js
+++ b/lib/test.js
@@ -211,18 +211,12 @@ Test.prototype.exit = function () {
 
 			self._checkPlanCount();
 
-			if (!self.ended) {
-				self.ended = true;
-
-				globals.setImmediate(function () {
-					if (self.assertError !== undefined) {
-						self.promise.reject(self.assertError);
-						return;
-					}
-
-					self.promise.resolve(self);
-				});
+			if (self.assertError !== undefined) {
+				self.promise.reject(self.assertError);
+				return;
 			}
+
+			self.promise.resolve(self);
 		});
 };