--- a/lib/test.js
+++ b/lib/test.js
@@ -101,12 +101,6 @@ Test.prototype.run = function () {
 		self.promise.reject = reject;
 	});
 
-	// TODO(vdemedes): refactor this to avoid storing the promise
-	if (!this.fn) {
-		this.exit();
-		return undefined;
-	}
-
 	this._timeStart = globals.now();
 
 	// wait until all assertions are complete