--- a/lib/test.js
+++ b/lib/test.js
@@ -228,16 +228,22 @@ function onAssertionEvent(event) {
 	var promise;
 	if (event.assertionThrew) {
 		event.error.powerAssertContext = event.powerAssertContext;
+		event.error.originalMessage = event.originalMessage;
 		promise = Promise.reject(event.error);
 	} else {
 		var ret = event.returnValue;
-		promise = isObservable(ret) ? observableToPromise(ret) : Promise.resolve(ret);
+		if (isObservable(ret)) {
+			ret = observableToPromise(ret);
+		}
+		if (isPromise(ret)) {
+			ret = ret
+				.then(null, function (err) {
+					err.originalMessage = event.originalMessage;
+					throw err;
+				});
+		}
+		promise = ret;
 	}
-	promise = promise
-		.catch(function (err) {
-			err.originalMessage = event.originalMessage;
-			return Promise.reject(err);
-		});
 	this._test._assert(promise);
 	return promise;
 }