--- a/package.json
+++ b/package.json
@@ -95,7 +95,7 @@
     "core-assert": "^0.1.0",
     "debug": "^2.2.0",
     "deeper": "^2.1.0",
-    "empower-core": "^0.4.0",
+    "empower-core": "github:jamestalmage/empower-core#enhance-prototypes",
     "figures": "^1.4.0",
     "find-cache-dir": "^0.1.1",
     "fn-name": "^2.0.0",