Using the [emoji-aware test-per-character branch](https://github.com/beaugunderson/emoji-aware/commit/e2c3a824ec6e28c2e86e9191fd7123cea8ac1e00) as a benchmark, and running `time ava`:


master:
```
  4363 passed

ava  22.61s user 0.54s system 109% cpu 21.093 total
```

this branch (only the `Runner` commits):
```
  4363 passed

ava  7.30s user 0.51s system 139% cpu 5.597 total
```

this branch (all the commits):
```
  4363 passed

ava  6.32s user 0.46s system 145% cpu 4.675 total
```

--- 

updated benchmark (with empower-core performance tweaks):

```
ava --verbose  5.05s user 0.32s system 138% cpu 3.872 total
```