@jamestalmage 2016-01-21
This relies on some changes some changes to `empower-core`.