@jamestalmage 2016-01-21
Further splitting that test suite into multiple files:

```
  4363 passed

ava  10.69s user 0.83s system 453% cpu 2.539 total
```