@sindresorhus 2016-01-21
:+1: Looks good! (ᵔᴥᵔ)

Too bad https://github.com/jamestalmage/ava/commit/4e23943eb9a54fb74ada303f3bfed81d7aade933 was slow. It was more readable before.