@jamestalmage 2016-01-21
> Too bad jamestalmage@4e23943 was slow. It was more readable.

Yeah, unfortunately it is a big offender in this scenario.