@jamestalmage 2016-01-21
Also a technically breaking change (though not one I am really concerned about):

```js
t.is(1 + 1, 2);
```

No longer returns a Promise. 

Only `t.throws` and `t.doesNotThrow` will return a promise.