@sindresorhus 2016-01-21
> though not one I am really concerned about

Me neither.