@sindresorhus 2016-01-21
`(つ◕౪◕)つ━☆ﾟ.*･｡ﾟ`