@jamestalmage 2016-01-22
@kentcdodds, et. all

Just for comparison - I also created a branch converting [`emoji-aware` to mocha](https://github.com/jamestalmage/emoji-aware/tree/mocha-test-per-char)

AVA now ekes out slightly better performance:

```
ava --verbose  9.76s user   0.71s system    465% cpu    2.252 total
mocha          2.07s user   0.31s system     96% cpu    2.460 total   
```

However, it's obvious the only reason AVA is winning is that it is leveraging multiple processes. Hopefully this means there are still lots of opportunities to make AVA even faster.